+++ Source/Krypton Components/Krypton.Toolkit/General/Accessibility/KryptonMaskedTextBoxAccessibleObject.cs
@@ -1,8 +1,8 @@
-#region BSD License
+﻿#region BSD License
 /*
  *
  *  New BSD 3-Clause License (https://github.com/Krypton-Suite/Standard-Toolkit/blob/master/LICENSE)
- *  Modifications by Peter Wagner (aka Wagnerp), Simon Coghlan (aka Smurf-IV), Giduac & Ahmed Abdelhameed et al. 2026 - 2026. All rights reserved.
+ *  Modifications by Peter Wagner (aka Wagnerp), Simon Coghlan (aka Smurf-IV), Giduac, Ahmed Abdelhameed, tobitege, KamaniAR, Lesandro Gotardo (aka lesandrog), Jorge A. Avilés (aka mcpbcs) et al. 2026 - 2026. All rights reserved.
  *
  */
 #endregion
@@ -39,11 +39,17 @@ internal class KryptonMaskedTextBoxAccessibleObject : Control.ControlAccessibleO
     {
         get
         {
+            if (!string.IsNullOrEmpty(_owner.AccessibleName))
+            {
+                return _owner.AccessibleName;
+            }
+
             // Try to get name from internal MaskedTextBox first
             var internalAccessible = _owner.MaskedTextBox?.AccessibilityObject;
-            if (internalAccessible?.Name != null)
+            string? name = internalAccessible?.Name;
+            if (!string.IsNullOrEmpty(name))
             {
-                return internalAccessible.Name;
+                return name;
             }
 
             // Fall back to base implementation
@@ -58,11 +64,17 @@ internal class KryptonMaskedTextBoxAccessibleObject : Control.ControlAccessibleO
     {
         get
         {
+            if (!string.IsNullOrEmpty(_owner.AccessibleDescription))
+            {
+                return _owner.AccessibleDescription;
+            }
+
             // Try to get description from internal MaskedTextBox first
             var internalAccessible = _owner.MaskedTextBox?.AccessibilityObject;
-            if (internalAccessible?.Description != null)
+            string? description = internalAccessible?.Description;
+            if (!string.IsNullOrEmpty(description))
             {
-                return internalAccessible.Description;
+                return description;
             }
 
             // Fall back to base implementation
@@ -129,6 +141,20 @@ internal class KryptonMaskedTextBoxAccessibleObject : Control.ControlAccessibleO
 
             // Fall back to control's text
             return _owner.Text;
+        }
+
+        set
+        {
+            string text = value ?? string.Empty;
+
+            if (_owner.MaskedTextBox != null)
+            {
+                _owner.MaskedTextBox.Text = text;
+            }
+            else
+            {
+                _owner.Text = text;
+            }
         }
     }